pyproject.toml
[project]
name = "bedrock-agents-defaults"
version = "0.1.0"
description = "Minimal Bedrock Agents repro + fix for the 'MANDATORY DEFAULTS' instruction-tuning bug on Claude Haiku 4.5"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0,<11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stacks"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
